requirements.txt
yfinance
pandas
python-dotenv
snowflake-connector-python
logging
snowflake-connector-python[pandas]